refactor: move TransactionPreviewService from model to service layer
